extract random event functions, need them for dispatcher
